Remove CH code from legacy class and remove material_view.sql
