--- src/main/java/org/cbioportal/persistence/mybatis/UnImplementedStudyViewMyBatisRepository.java
@@ -1,198 +1,0 @@
-package org.cbioportal.persistence.mybatis;
-
-import org.cbioportal.model.AlterationCountByGene;
-import org.cbioportal.model.CaseListDataCount;
-import org.cbioportal.model.ClinicalAttribute;
-import org.cbioportal.model.ClinicalData;
-import org.cbioportal.model.ClinicalDataCount;
-import org.cbioportal.model.ClinicalDataCountItem;
-import org.cbioportal.model.ClinicalEventTypeCount;
-import org.cbioportal.model.CopyNumberCountByGene;
-import org.cbioportal.model.GenericAssayDataCountItem;
-import org.cbioportal.model.GenomicDataCount;
-import org.cbioportal.model.GenomicDataCountItem;
-import org.cbioportal.model.MolecularProfile;
-import org.cbioportal.model.PatientTreatment;
-import org.cbioportal.model.Sample;
-import org.cbioportal.model.SampleTreatment;
-import org.cbioportal.model.StudyViewFilterContext;
-import org.cbioportal.persistence.StudyViewRepository;
-import org.cbioportal.utils.config.annotation.ConditionalOnProperty;
-import org.cbioportal.web.parameter.ClinicalDataType;
-import org.cbioportal.web.parameter.GenericAssayDataBinFilter;
-import org.cbioportal.web.parameter.GenericAssayDataFilter;
-import org.cbioportal.web.parameter.GenomicDataBinFilter;
-import org.cbioportal.web.parameter.GenomicDataFilter;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-@Repository
-@ConditionalOnProperty(name = "clickhouse_mode", havingValue = "false", matchIfMissing = true)
-public class UnImplementedStudyViewMyBatisRepository implements StudyViewRepository {
-    
-    private static final String UNSUPPORTED_OPERATION = "StudyViewRepository Feature Not supported...  Contact cbio admin.";
-    
-    @Override
-    public List<Sample> getFilteredSamples(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<String> getFilteredStudyIds(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<ClinicalData> getSampleClinicalData(StudyViewFilterContext studyViewFilterContext, List<String> attributeIds) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<ClinicalData> getPatientClinicalData(StudyViewFilterContext studyViewFilterContext, List<String> attributeIds) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-
-    }
-
-    @Override
-    public List<AlterationCountByGene> getMutatedGenes(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<AlterationCountByGene> getStructuralVariantGenes(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<CopyNumberCountByGene> getCnaGenes(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<ClinicalDataCountItem> getClinicalDataCounts(StudyViewFilterContext studyViewFilterContext, List<String> filteredAttributes) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<GenomicDataCount> getMolecularProfileSampleCounts(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public Map<String, ClinicalDataType> getClinicalAttributeDatatypeMap() {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<CaseListDataCount> getCaseListDataCountsPerStudy(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public Map<String, Integer> getTotalProfiledCounts(StudyViewFilterContext studyViewFilterContext, String alterationType, List<MolecularProfile> molecularProfiles) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<ClinicalAttribute> getClinicalAttributes() {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<ClinicalAttribute> getClinicalAttributesForStudies(List<String> studyIds) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public int getFilteredSamplesCount(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public int getFilteredPatientCount(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public Map<String, Set<String>> getMatchingGenePanelIds(StudyViewFilterContext studyViewFilterContext, String alterationType) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public int getTotalProfiledCountsByAlterationType(StudyViewFilterContext studyViewFilterContext, String alterationType) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public int getSampleProfileCountWithoutPanelData(StudyViewFilterContext studyViewFilterContext, String alterationType) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<ClinicalEventTypeCount> getClinicalEventTypeCounts(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<PatientTreatment> getPatientTreatments(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public int getTotalPatientTreatmentCount(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<SampleTreatment> getSampleTreatments(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public int getTotalSampleTreatmentCount(StudyViewFilterContext studyViewFilterContext) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<GenomicDataCountItem> getCNACounts(StudyViewFilterContext studyViewFilterContext, List<GenomicDataFilter> genomicDataFilters) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<GenericAssayDataCountItem> getGenericAssayDataCounts(StudyViewFilterContext studyViewFilterContext, List<GenericAssayDataFilter> genericAssayDataFilters) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public Map<String, Integer> getMutationCounts(StudyViewFilterContext studyViewFilterContext, GenomicDataFilter genomicDataFilter) {
-        throw new UnsupportedOperationException(UNSUPPORTED_OPERATION);
-    }
-
-    @Override
-    public List<GenomicDataCountItem> getMutationCountsByType(StudyViewFilterContext studyViewFilterContext, List<GenomicDataFilter> genomicDataFilters) {
-        return List.of();
-    }
-
-    @Override
-    public List<ClinicalDataCount> getGenomicDataBinCounts(StudyViewFilterContext studyViewFilterContext, List<GenomicDataBinFilter> genomicDataBinFilters) {
-        return List.of();
-    }
-
-    @Override
-    public List<ClinicalDataCount> getGenericAssayDataBinCounts(StudyViewFilterContext studyViewFilterContext, List<GenericAssayDataBinFilter> genericAssayDataBinFilters) {
-        return List.of();
-    }
-
-    @Override
-    public List<MolecularProfile> getGenericAssayProfiles() {
-        return List.of();
-    }
-
-    @Override
-    public List<MolecularProfile> getFilteredMolecularProfilesByAlterationType(StudyViewFilterContext studyViewFilterContext, String alterationType) {
-        return List.of();
-    }
-}
